Remove the dependency on uncompyle6
It turns out that uncompyle6 added a GPL dependency that I was
previously unaware of because it was hidden inside of xdis, which
is itself MIT licensed and didn't publish this fact.

GPL compliance is quite an onerous requirement to add for the sake
of pretty printing some functions, so I'm removing it! This goes
back to the old way of extracting lambda sources, which is less
effective but works in most cases.